pyproject.toml
[tool.poetry]
name = "quickcontract"
version = "0.1.0"
description = "契約文生成API - URL から主要テキストを抽出し、Gemini で契約文を生成"
authors = ["QuickContract Team <team@quickcontract.dev>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
google-genai = "^0.8.0"
requests = "^2.31.0"
readability-lxml = "^0.8.1"
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[tool.poetry.scripts]
start = "uvicorn app.main:app --reload --host 0.0.0.0 --port 8000"
dev = "uvicorn app.main:app --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88